requirements.txt
requests
beautifulsoup4
pandas
pyyaml 
transformers
torch 
sentencepiece 